Creates an second version of the documentation, for the plugin versions that predates UE 5.3.2. Add a new "hidden" tag so the old one does not appear in the guides section.
The major changes to the plugin meant that post 5.3.2 some of the guide was deprecated. Separating the guide into two versions is meant to make this clearer. The old guide does not show up in the Guides menu, but has a new "hidden" tag meaning that it is only accessable through a link in the new guide.